Fix the shading in scud. When lighting is disabled the fixed shading intensities appear to be offset values added to the colour value.

+++ Src/Graphics/New3D/Model.h
@@ -16,7 +16,7 @@ struct Vertex
 	float pos[3];
 	float normal[3];
 	float texcoords[2];
-	UINT8 color[4];			//rgba
+	float color[4];			//rgba
 };
 
 struct Poly		// our polys are always 3 triangles, unlike the real h/w
